Shared model code reorganization: Separate `SharedModelManager` from `SharedModel` Separate shared model registration from tool registration Move shared model-related files from `models/tools` folder to `models/shared` A bunch of trailing spaces removed automatically due to VSCode setting
