+++ src/models/document/document.ts
@@ -10,8 +10,8 @@ import {
 import { AppConfigModelType } from "../stores/app-config-model";
 import { TileCommentsModel, TileCommentsModelType } from "../tools/tile-comments";
 import { UserStarModel, UserStarModelType } from "../tools/user-star";
-import { 
-  IDocumentMetadata, IGetNetworkDocumentParams, IGetNetworkDocumentResponse, IUserContext 
+import {
+  IDocumentMetadata, IGetNetworkDocumentParams, IGetNetworkDocumentResponse, IUserContext
 } from "../../../functions/src/shared";
 import { getFirebaseFunction } from "../../hooks/use-firebase-function";
 import { IDocumentProperties } from "../../lib/db-types";
@@ -19,7 +19,7 @@ import { getLocalTimeStamp } from "../../utilities/time";
 import { safeJsonParse } from "../../utilities/js-utils";
 import { Tree } from "../history/tree";
 import { TreeMonitor } from "../history/tree-monitor";
-import { ISharedModelDocumentManager, SharedModelDocumentManager } from "../tools/shared-model-document-manager";
+import { ISharedModelDocumentManager, SharedModelDocumentManager } from "../shared/shared-model-document-manager";
 import { ITileEnvironment } from "../tools/tool-types";
 import { TreeManager } from "../history/tree-manager";
 import { ESupportType } from "../curriculum/support";
@@ -103,7 +103,7 @@ export const DocumentModel = Tree.named("Document")
       // contextId. However the old functions do. The old functions were just
       // ignoring this contextId. So the contextId is added here so the client
       // code can work with the old functions.
-      return { contextId: "ignored", uid, type, key, createdAt, title, 
+      return { contextId: "ignored", uid, type, key, createdAt, title,
         originDoc, properties: properties.toJSON() } as IDocumentMetadata;
     },
     getProperty(key: string) {
@@ -381,7 +381,7 @@ export const createDocumentModel = (snapshot?: DocumentModelSnapshotType) => {
       throw e;
     }
 
-    // Putting the error in an object like this prevents Chrome from expanding the 
+    // Putting the error in an object like this prevents Chrome from expanding the
     // error and taking up a bunch of console lines.
     console.error("Failed to load document", {docKey: snapshot.key, error: e});
