+++ src/models/tools/text/text-plugin-info.ts
@@ -1,16 +1,16 @@
 import { HtmlSerializablePlugin } from "@concord-consortium/slate-editor";
 import { FunctionComponent, SVGProps } from "react";
-import { SharedModelType } from "../shared-model";
+import { SharedModelType } from "../../shared/shared-model";
 import { TextContentModelType } from "./text-content";
 
 export interface ITextPluginInfo {
   iconName: string;
   Icon: FunctionComponent<SVGProps<SVGSVGElement>>;
   toolTip: string;
-  createSlatePlugin: 
+  createSlatePlugin:
     (textContent: TextContentModelType) => HtmlSerializablePlugin;
   command: string;
-  updateTextContentAfterSharedModelChanges?: 
+  updateTextContentAfterSharedModelChanges?:
     (textContent: TextContentModelType, sharedModel?: SharedModelType) => void;
 }
 
@@ -25,7 +25,7 @@ export function getTextPluginInfo(id: string) {
 }
 
 // TODO: perhaps this should only add the plugins that have been configured
-// as tools by the app-config. 
+// as tools by the app-config.
 export function getTextPluginInstances(textContent: TextContentModelType) {
   const pluginInstances:  HtmlSerializablePlugin[] = [];
   Object.values(gTextPluginInfoMap).forEach(pluginInfo => {
