+++ src/models/tools/tool-content-info.ts
@@ -2,7 +2,6 @@ import { FunctionComponent, SVGProps } from "react";
 import { ToolContentModel, ToolContentModelType, ToolMetadataModel } from "./tool-types";
 import { IToolTileProps } from "../../components/tools/tool-tile";
 import { AppConfigModelType } from "../stores/app-config-model";
-import { SharedModel } from "./shared-model";
 
 export interface IDMap {
   [id: string]: string;
@@ -82,23 +81,4 @@ export interface IDocumentExportOptions extends ITileExportOptions {
   includeTileIds?: boolean;
   appendComma?: boolean;
   transformImageUrl?: (url: string, filename?: string) => string;
-}
-
-export interface ISharedModelInfo {
-  type: string;
-  modelClass: typeof SharedModel;
-}
-
-const gSharedModelInfoMap: Record<string, ISharedModelInfo> = {};
-
-export function registerSharedModelInfo(sharedModelInfo: ISharedModelInfo) {
-  gSharedModelInfoMap[sharedModelInfo.type] = sharedModelInfo;
-}
-
-export function getSharedModelClasses() {
-  return Object.values(gSharedModelInfoMap).map(info => info.modelClass);
-}
-
-export function getSharedModelInfoByType(type: string) {
-  return type ? gSharedModelInfoMap[type] : undefined;
 }
