+++ src/models/tools/tool-types.ts
@@ -1,5 +1,6 @@
 import { getEnv, getSnapshot, Instance, ISerializedActionCall, types } from "mobx-state-tree";
-import { ISharedModelManager, SharedModelType } from "./shared-model";
+import { SharedModelType } from "../shared/shared-model";
+import { ISharedModelManager } from "../shared/shared-model-manager";
 import { getToolContentModels, getToolContentInfoById } from "./tool-content-info";
 
 /**
@@ -46,7 +47,7 @@ export interface IToolContentModelHooks {
   /**
    * This is called before the tile is removed from the row of the document.
    * Immediately after the tile is removed from the row it is also removed from
-   * the tileMap which is the actual container of the tile. 
+   * the tileMap which is the actual container of the tile.
    */
   willRemoveFromDocument?(): void
 }
@@ -118,18 +119,18 @@ export const ToolContentModel = types.model("ToolContentModel", {
   }))
   .actions(self => ({
     /**
-     * This will be called automatically by the tree monitor. 
+     * This will be called automatically by the tree monitor.
      * Currently the call tree looks like:
      * TreeMonitor.recordAction
      * └ Tree.handleSharedModelChanges
      *   └ Tree.updateTreeAfterSharedModelChangesInternal
      *     └ Tree.updateTreeAfterSharedModelChanges
      *       └ tile.content.updateAfterSharedModelChanges
-     * 
+     *
      * It is also called after the manager has finished applying patches
      * during an undo or replying history.
-     * 
-     * @param sharedModel 
+     *
+     * @param sharedModel
      */
     updateAfterSharedModelChanges(sharedModel?: SharedModelType) {
       throw new Error("not implemented");
