+++ src/plugins/shared-variables/shared-variables-registration.ts
@@ -1,11 +1,11 @@
+import { registerSharedModelInfo } from "../../models/shared/shared-model-registry";
 import { registerTextPluginInfo } from "../../models/tools/text/text-plugin-info";
-import { registerSharedModelInfo } from "../../models/tools/tool-content-info";
 import { kSharedVariablesID, SharedVariables } from "./shared-variables";
 import VariablesToolIcon from "./slate/variables.svg";
 import { VariablesPlugin } from "./slate/variables-plugin";
 import { updateAfterSharedModelChanges } from "./slate/variables-text-content";
 import { registerDrawingObjectInfo, registerDrawingToolInfo } from "../drawing-tool/components/drawing-object-manager";
-import { EditVariableButton, EditVariableTool, VariableChipComponent, VariableChipObject, VariableChipToolbarButton, 
+import { EditVariableButton, EditVariableTool, VariableChipComponent, VariableChipObject, VariableChipToolbarButton,
   VariableDrawingTool } from "./drawing/variable-object";
 
 registerSharedModelInfo({
@@ -39,5 +39,3 @@ registerDrawingToolInfo({
   toolClass: VariableDrawingTool,
   buttonComponent: VariableChipToolbarButton
 });
-
-
