+++ src/plugins/shared-variables/slate/variables-text-content.test.tsx
@@ -2,7 +2,8 @@ import React from "react";
 import { IAnyStateTreeNode, IAnyType, types, castToSnapshot } from "mobx-state-tree";
 import { Editor, SlateEditor } from "@concord-consortium/slate-editor";
 import { render } from "@testing-library/react";
-import { ISharedModelManager, SharedModelType } from "../../../models/tools/shared-model";
+import { SharedModelType } from "../../../models/shared/shared-model";
+import { ISharedModelManager } from "../../../models/shared/shared-model-manager";
 import { TextContentModel, TextContentModelType } from "../../../models/tools/text/text-content";
 import { SharedVariables, SharedVariablesType } from "../shared-variables";
 import { getOrFindSharedModel, updateAfterSharedModelChanges } from "./variables-text-content";
